backend/requirements.txt
fastapi 
uvicorn 
tensorflow 
pillow 
python-multipart
firebase-admin
cryptography